% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basics.R
\name{add.scale.bar}
\alias{add.scale.bar}
\title{Add scale bar to diagonal heatmap}
\usage{
add.scale.bar(
  bin.size,
  units,
  x.at = 0.8,
  y.at = 0.8,
  width = 0.2,
  lwd = 3,
  label.vadj = 0.025
)
}
\arguments{
\item{bin.size}{Width of each bin}

\item{units}{Suffix for scale bar (default: "kb")}

\item{x.at}{Relative position on x-axis for center of scale bar (default: 0.8)}

\item{y.at}{Relative position on y-axis for center of scale bar (default: 0.8)}

\item{width}{Relative width of scale bar (default: 0.2)}

\item{lwd}{Line width for scale bar (default: 3)}

\item{label.vadj}{Relative vertical adjustment factor for label spacing (default: 0.025)}
}
\description{
Add a 1D scale bar to a diagonal heatmap produced by \code{\link[=plot.diag.heat]{plot.diag.heat()}}
}
\details{
\code{width} will be rounded to the nearest whole value of bin.size
}
\examples{
plot.diag.heat(matrix(rnorm(100), nrow=10))
add.scale.bar(bin.size=1000, units="kb", label.vadj=-0.01)

}
\seealso{
\code{\link[=plot.diag.heat]{plot.diag.heat()}}
}
