% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants}
\alias{get_constants}
\title{Load DSMap constants}
\usage{
get_constants(subset)
}
\arguments{
\item{susbet}{Name of constant subset to return. See \code{Details} for options.}
}
\value{
Returns a vector of constants corresponding to the value of \code{subset}
}
\description{
Load a subset of constants for DSMap analyses
}
\details{
Recognized values for \code{subset} include:
\itemize{
\item \code{colors} : all color palettes used
\item \code{cnv_colors} : color palettes for CNVs (subset of \code{colors})
\item \code{other_colors} : all other color constants not included in CNV palettes (subset of \code{colors})
}
}
\examples{
# Load list of color palettes
dsmap.colors <- get_constants("colors");
}
