% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_pair_utils.R
\name{infer.bin.size}
\alias{infer.bin.size}
\title{Infer bin size}
\usage{
infer.bin.size(coords, n_samp = 1000)
}
\arguments{
\item{coords}{Dataframe styled as BED (first three columns are coordinates)}

\item{n_samp}{number of rows to sample from dataframe. Default = 1,000.}
}
\value{
Integer estimate of inferred bin size
}
\description{
Infer bin size from BED file loaded as dataframe
}
\examples{
# Load coordinates from a BED file & infer bin size
bins <- dsmapR::load.bins("path/to/my/bins.bed")
binsize <- infer.bin.size(bins$coords)

}
