% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_pair_utils.R
\name{load.bins}
\alias{load.bins}
\title{Load athena bins or bin-pairs}
\usage{
load.bins(bed_in, n_keep_features = Inf, feats_are_numeric = TRUE)
}
\arguments{
\item{bed_in}{Path to input BED file}

\item{n_keep_features}{Number of feature columns to keep. Defaults to keeping
all feature columns present in \code{bed_in}.}

\item{feats_are_numeric}{Boolean indicating if all features should be coerced
to numeric type. \code{TRUE} by default.}
}
\value{
Returns a list of two dataframes:
\enumerate{
\item \verb{$coords}: a three-column dataframe of bin coordinates
\item \verb{$feats}: a dataframe of bin features
}
}
\description{
Load a BED file of athena 1D bins or 2D bin-pairs
}
