% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{load.constants}
\alias{load.constants}
\title{Load DSMap constants}
\usage{
load.constants(subset)
}
\arguments{
\item{subset}{Vector of constant groups to load See \code{Details} for options.}
}
\description{
Load a subset of constants for DSMap analyses
}
\details{
Recognized values for \code{subset} include:
\itemize{
\item \code{colors} : all color palettes used
\item \code{cnv_colors} : color palettes for CNVs (subset of \code{colors})
\item \code{other_colors} : all other color constants not included in CNV palettes (subset of \code{colors})
\item \code{scales} : all scales and scale labels
\item \code{names} : names of various variables
\item \code{all} : load all constants
}
}
\examples{
# Load list of color palettes
get.constants("colors");

# Load scales and CNV colors
get.constants(c("scales", "cnv_colors"))

}
