% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapEncadrement.R
\name{makeHeatmapEncadre}
\alias{makeHeatmapEncadre}
\title{makeHeatmapEncadre}
\usage{
makeHeatmapEncadre(
  matrice,
  VecIndStruc,
  LisTMS,
  supcol = "green",
  infcol = "red",
  backgroundcol = "black",
  Legendname = "Legend",
  color = viridis::magma(256),
  width = NULL,
  column_names_max_height = max_text_width(colnames(matrice)),
  column_names_gp = gpar(fontsize = 0.2 + 1/log10(ncol(matrice)), col = colaff),
  row_names_gp = gpar(fontsize = 0.2 + 1/log10(nrow(matrice)), col = colaff),
  show_row_names = FALSE,
  colorBorder = "white",
  row_title = NULL,
  row_gap = unit(0, "mm"),
  column_title = NULL,
  column_gap = unit(0, "mm"),
  legend_param = list(title_gp = gpar(col = colaff), labels_gp = gpar(col = colaff)),
  AnnoClustCol = rainbow,
  AnnoLineCol = "yellow",
  colPsup = "chocolate1",
  colPinf = "turquoise1"
)
}
\arguments{
\item{matrice}{une matrice}

\item{VecIndStruc}{un vecteur issu d'une methode d'indice de structuration}

\item{LisTMS}{la liste resultant de la fonction TransMeanSeuil}

\item{supcol}{couleur si moyenne superieur au seuil}

\item{infcol}{couleur si moyenne inferieur au seuil}

\item{backgroundcol}{couleur de fond "black" ou "white"}

\item{Legendname}{nom de la legende}

\item{color}{palette de couleur}

\item{width}{largeur de la heatmap}

\item{column_names_max_height}{voir Heatmap}

\item{column_names_gp}{voir Heatmap}

\item{row_names_gp}{voir Heatmap}

\item{show_row_names}{voir Heatmap}

\item{colorBorder}{couleur des barrieres sur la heatmap}

\item{row_title}{voir Heatmap}

\item{row_gap}{voir Heatmap}

\item{column_title}{voir Heatmap}

\item{column_gap}{voir Heatmap}

\item{legend_param}{voir Heatmap}

\item{AnnoClustCol}{une palette de couleur}

\item{AnnoLineCol}{une couleur pour faire les lignes des seuils}

\item{colPsup}{couleur si point superieur au seuil}

\item{colPinf}{couleur si point inferieur au seuil}
}
\description{
makeHeatmapEncadre
}
