% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basics.R
\name{plot.diag.heat}
\alias{plot.diag.heat}
\title{Plot diagonal heatmap}
\usage{
\method{plot}{diag.heat}(
  mat,
  orient = "up",
  gridlines = TRUE,
  gridline.lwd = 0.25,
  parmar = c(0.5, 0.5, 0.5, 0.5)
)
}
\arguments{
\item{mat}{Numeric square matrix (or data.frame) of values to be plotted}

\item{orient}{Specify orientation of plot as "up" diagonal pointing up or
"down" for diagonal pointing down (default: "up")}

\item{gridlines}{Boolean indicator whether to add gridlines (default: true)}

\item{gridline.lwd}{Width of gridlines (default: 0.1)}

\item{parmar}{Margin values passed to par()}
}
\description{
Plot the upper-right diagonal of a numeric square matrix as a rotated heatmap
}
\seealso{
\code{\link[=prep.plot.area]{prep.plot.area()}}, \code{\link[=rotate.points]{rotate.points()}}
}
