% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basics.R
\name{prep_plot_area}
\alias{prep_plot_area}
\title{Prepare plot area}
\usage{
prep_plot_area(xlims, ylims, parmar)
}
\arguments{
\item{xlims}{Range of values for X axis}

\item{ylims}{Range of values for Y axis}

\item{parmar}{Margin values passed to par()}
}
\description{
Prepare a standardized & formatted plot area
}
\examples{
prep_plot_area(xlims=c(0, 5), ylims=(-10, 10), parmar=rep(3, 4));
}
