% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basics.R
\name{rotate.points}
\alias{rotate.points}
\title{Rotate points}
\usage{
rotate.points(coords, angle, x.origin = 0, y.origin = 0)
}
\arguments{
\item{coords}{Two-column matrix or data.frame of original x & y coordinates}

\item{angle}{Angle to rotate, in degrees}

\item{x.origin}{X coordinate to rotate about (default: 0)}

\item{y.origin}{Y coordinate to rotate about (default: 0)}
}
\value{
Returns a two-column data.frame of rotated x & y coordinates
}
\description{
Rotate one or more points in 2D Cartesian space
}
\examples{
rotate.points(coords=data.frame("x"=c(2, 5), "y"=c(0, 3)), angle=90)
}
