% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransMeanSeuil.R
\name{TransMeanSeuil}
\alias{TransMeanSeuil}
\title{TransMeanSeuil}
\usage{
TransMeanSeuil(
  VecIndStruc,
  Ssup = NULL,
  Sinf = NULL,
  methodSplit = "transition",
  GrouporSeuil = "group",
  Pval = 0.05,
  minsize = 3,
  conf.level = 0.95
)
}
\arguments{
\item{VecIndStruc}{un vecteur issu d'une methode d'indice de structuration}

\item{Ssup}{une valeur seuil superieur}

\item{Sinf}{une valeur seuil inferieur}

\item{methodSplit}{chaine de caractere "transition" ou "mean"}

\item{GrouporSeuil}{chaine de caractere "group" ou "seuil"}

\item{Pval}{une p-value pour la fonction e.divisive du package ecp}

\item{minsize}{un nombre minimum d'observation pour la fonction e.divisive du package ecp}

\item{conf.level}{un niveau de confiance pour le test de student}
}
\value{
une liste contenant
}
\description{
TransMeanSeuil
}
