% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecARorder.R
\name{vecARorder}
\alias{vecARorder}
\title{vecARorder}
\usage{
vecARorder(dist, seriation)
}
\arguments{
\item{dist}{un objet de type dist}

\item{seriation}{une seriation pour organiser la matrice de distance}
}
\value{
un vecteur
}
\description{
vecARorder
}
