% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecCorrOrder.R
\name{vecCorrOrder}
\alias{vecCorrOrder}
\title{vecCorrOrder}
\usage{
vecCorrOrder(dist, seriation)
}
\arguments{
\item{dist}{un objet de type dist}

\item{seriation}{une seriation pour organiser la matrice de distance}
}
\value{
un vecteur
}
\description{
vecCorrOrder
}
