% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecECPrunTest.R
\name{vecECPrunTest}
\alias{vecECPrunTest}
\title{vecECPrunTest}
\usage{
vecECPrunTest(matrice, pval = 0.01, N = 3)
}
\arguments{
\item{matrice}{une matrice}

\item{pval}{valeur de p-value}

\item{N}{nombre minimum d'observation, voir 'min.size' de la fonction e.divisive du package ecp}
}
\value{
un vecteur
}
\description{
vecECPrunTest
}
