% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecRunsTest.R
\name{vecRunsTest}
\alias{vecRunsTest}
\title{vecRunsTest}
\usage{
vecRunsTest(matrice, alternative = "two.sided")
}
\arguments{
\item{matrice}{une matrice}

\item{alternative}{voir 'alternative' de la fonction e.divisive du package ecp}
}
\value{
un vecteur
}
\description{
vecRunsTest
}
